House Of Young/requirements.txt
asgiref==3.7.2
Django==4.2.9
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.14.0
MarkupSafe==2.1.3
pillow==10.2.0
pypng==0.20220715.0
python-decouple==3.8
pytz==2023.3.post1
qrcode==7.4.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
Werkzeug==3.0.1
whitenoise==6.6.0
